[New version big refactor]: Refine service and setup for new version
