+++ components/common/index.ts
@@ -1,3 +1,4 @@
 export { default as Layout } from './Layout'
 export { default as NavBar } from './Navbar'
+export { default as Footer } from './Footer'
 export { default as Head } from './Head'
