+++ components/common/user/index.tsx
@@ -1,0 +1,9 @@
+export { default as CreateAccountPage } from './CreateAccountPage'
+export { default as SuccessfullyRegisteredPage } from './SuccessfullyRegisteredPage'
+
+export { default as VerifyEmailPage } from './VerifyEmailPage'
+export { default as CompleteProfilePage } from './CompleteProfilePage'
+
+export { default as LoginPage } from './LoginPage'
+export { default as ChangePasswordPage } from './ChangePasswordPage'
+export { default as RequestNewPasswordPage } from './RequestNewPasswordPage'
