+++ components/mentee/index.ts
@@ -1,0 +1,1 @@
+export { default as MenteeProfileView } from './ProfileView'
