+++ components/mentor/ProfilePage.tsx
@@ -1,0 +1,1 @@
+export default {}
