+++ components/mentor/index.ts
@@ -1,0 +1,2 @@
+export { default as MentorProfilePage } from './ProfilePage'
+export { default as MentorCard } from './MentorCard'
