+++ components/ui/Benefits/Benefits.tsx
@@ -1,0 +1,5 @@
+const Benefits = () => {
+  return <div>Benefits</div>
+}
+
+export default Benefits
