+++ components/ui/Feedback/index.tsx
@@ -1,0 +1,5 @@
+const Feedback = () => {
+  return <div>Feedback</div>
+}
+
+export default Feedback
