+++ components/ui/Hero/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Hero'
