+++ components/ui/Values/index.tsx
@@ -1,0 +1,5 @@
+const Values = () => {
+  return <div>Values</div>
+}
+
+export default Values
