+++ components/ui/index.ts
@@ -3,6 +3,12 @@ export { default as Checkbox } from './Checkbox'
 export { default as Input } from './Input'
 export { default as Link } from './Link'
 export { default as Select } from './Select'
-export { default as Typography } from './Typography'
+
 export { default as Theme } from './Theme'
+export { default as Typography } from './Typography'
 export { default as Skeleton } from './Skeleton'
+
+export { default as Hero } from './Hero'
+export { default as Benefits } from './Benefits'
+export { default as Values } from './Values'
+export { default as Feedback } from './Feedback'
