--- components/user/index.tsx
@@ -1,8 +1,0 @@
-export { default as CreateAccount } from './CreateAccountView'
-
-export { default as VerifyEmail } from './VerifyEmailView'
-export { default as CompleteProfile } from './CompleteProfileView'
-
-export { default as LoginView } from './LoginView'
-export { default as ChangePassword } from './ChangePasswordView'
-export { default as RequestNewPassword } from './RequestNewPasswordView'
