+++ pages/user/create-account.tsx
@@ -1,16 +1,3 @@
-import Head from 'next/head'
-import { CreateAccount } from '~/components/user'
-
-const CreateAccountPage = () => {
-  return (
-    <>
-      <Head>
-        <title>Create New Account</title>
-      </Head>
-
-      <CreateAccount />
-    </>
-  )
-}
+import { CreateAccountPage } from '~/components/common/user'
 
 export default CreateAccountPage
