+++ pages/user/successfully-registered.tsx
@@ -1,0 +1,3 @@
+import { SuccessfullyRegisteredPage } from '~/components/common/user'
+
+export default SuccessfullyRegisteredPage
